+++ src/server/formations/infra/repositories/apiTrajectoiresProStatistique.repository.ts
@@ -1,36 +1,53 @@
-import { createFailure, createSuccess, Either, isFailure } from '~/server/errors/either';
-import { ErreurMétier } from '~/server/errors/erreurMétier.types';
+import { createSuccess, Either, isFailure } from '~/server/errors/either';
 import { Statistique } from '~/server/formations/domain/statistique';
 import { StatistiqueRepository } from '~/server/formations/domain/statistique.repository';
 import {
 	ApiTrajectoiresProStatistiqueResponse,
-	isRegionEtAuMoinsUnPourcentageDisponible,
+	StatistiqueAvecRegionEtAuMoinsUnPourcentage,
+	StatistiquesMappedFromApi,
 } from '~/server/formations/infra/repositories/apiTrajectoiresProStatistique';
 import { mapStatistiques } from '~/server/formations/infra/repositories/apiTrajectoiresProStatistique.mapper';
-import { handleFailureError } from '~/server/formations/infra/repositories/apiTrajectoiresProStatistiqueError';
-import { ApiGeoRepository } from '~/server/localisations/infra/repositories/apiGeo.repository';
+import { LocalisationRepository } from '~/server/localisations/domain/localisation.repository';
+import { ErrorManagementService } from '~/server/services/error/errorManagement.service';
 import { PublicHttpClientService } from '~/server/services/http/publicHttpClient.service';
-import { LoggerService } from '~/server/services/logger.service';
 
 export class ApiTrajectoiresProStatistiqueRepository implements StatistiqueRepository {
-	constructor(private httpClientService: PublicHttpClientService, private apiGeoLocalisationRepository: ApiGeoRepository, private loggerService: LoggerService) {}
+	constructor(private readonly httpClientService: PublicHttpClientService, private readonly apiGeoLocalisationRepository: LocalisationRepository, private readonly errorManagementService: ErrorManagementService) {}
 
 	async get(codeCertification: string, codePostal: string): Promise<Either<Statistique>> {
 		try {
-			const codeRegion = await this.apiGeoLocalisationRepository.getCodeRegionByCodePostal(codePostal);
-			if (isFailure(codeRegion) || !codeRegion.result) {
-				return createFailure(ErreurMétier.SERVICE_INDISPONIBLE);
+			const codeRegionOrFailure = await this.apiGeoLocalisationRepository.getCodeRegionByCodePostal(codePostal);
+			if (isFailure(codeRegionOrFailure)) {
+				return codeRegionOrFailure;
 			}
 
 			const { data } = await this.httpClientService.get<ApiTrajectoiresProStatistiqueResponse>(
-				`inserjeunes/regionales/${codeRegion.result}/certifications/${codeCertification}`,
+				`inserjeunes/regionales/${codeRegionOrFailure.result}/certifications/${codeCertification}`,
 			);
 			const statistiques = mapStatistiques(data);
-			if (isRegionEtAuMoinsUnPourcentageDisponible(statistiques)) return createSuccess(statistiques);
-			return createFailure(ErreurMétier.CONTENU_INDISPONIBLE);
-		} catch (e) {
-			return handleFailureError(e, 'statistique formation', this.loggerService);
+			if (ApiTrajectoiresProStatistiqueRepository.isRegionEtAuMoinsUnPourcentageDisponible(statistiques)) {
+				return createSuccess(statistiques);
+			}
+
+			const incompleteStatistiqueError = new Error(JSON.stringify(statistiques));
+			return this.errorManagementService.handleFailureError(incompleteStatistiqueError, {
+				apiSource: 'API Trajectoires Pro',
+				contexte: 'get statistique de formation',
+				message: 'statistique de formation trouvée mais incomplète',
+			});
+		} catch (error) {
+			return this.errorManagementService.handleFailureError(error, {
+				apiSource: 'API Trajectoires Pro',
+				contexte: 'get statistique de formation',
+				message: 'statistique de formation non trouvée',
+			});
 		}
+	}
+
+	private static isRegionEtAuMoinsUnPourcentageDisponible(statistiquesMappedFromApi: StatistiquesMappedFromApi): statistiquesMappedFromApi is StatistiqueAvecRegionEtAuMoinsUnPourcentage {
+		const isRegionStatistiqueDisponible = !!statistiquesMappedFromApi.region;
+		const isStatistiqueDisponible = !!statistiquesMappedFromApi.tauxEnEmploi6Mois || !!statistiquesMappedFromApi.tauxEnFormation || !!statistiquesMappedFromApi.tauxAutres6Mois;
+		return isRegionStatistiqueDisponible && isStatistiqueDisponible;
 	}
 
 }